# src/app/swagger/api.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Pingr-- API",
    "description": "Pingr-- API Documentation"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/accounts": {
      "post": {
        "summary": "Creates a new account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "username",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "name": {
                    "type": "string"
                  },
                  "bio": {
                    "type": "string"
                  },
                  "birthDate": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created - Account"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/accounts/{id}": {
      "get": {
        "summary": "Gets account by Id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Account"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "summary": "Deletes account by Id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "session"
                ],
                "properties": {
                  "session": {
                    "type": "string",
                    "format": "session"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/accounts/{id}/auth": {
      "get": {
        "summary": "Checks if a specific account is logged",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Boolean"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Logs in a specific account",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK - Session Id"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/accounts/logout": {
      "post": {
        "summary": "Logs out a specific account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "session"
                ],
                "properties": {
                  "session": {
                    "type": "string",
                    "format": "session"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/{sessionId}": {
      "get": {
        "summary": "Validates and gets session information",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Account Id associated with the session"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/profiles": {
      "get": {
        "summary": "Gets profiles by query match",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string"
          },
          {
            "in": "query",
            "name": "username",
            "type": "string"
          },
          {
            "in": "query",
            "name": "bio",
            "type": "string"
          },
          {
            "in": "query",
            "name": "name",
            "type": "string"
          },
          {
            "in": "query",
            "name": "birthDate",
            "type": "date"
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Array of Profiles"
          }
        }
      }
    },
    "/profiles/usernames": {
      "get": {
        "summary": "Gets profiles by username prefix",
        "parameters": [
          {
            "in": "query",
            "name": "username",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Profiles"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/profiles/{id}": {
      "get": {
        "summary": "Gets profile by Id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Profile"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "summary": "Updates profile by Id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "session"
                ],
                "properties": {
                  "session": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "bio": {
                    "type": "string"
                  },
                  "birthDate": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK - Account"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/profiles/{id}/relationship": {
      "post": {
        "summary": "Follows or unfollows a specific profile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "session"
                ],
                "properties": {
                  "session": {
                    "type": "string",
                    "format": "session"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK - Action played"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/profiles/{id}/block": {
      "post": {
        "summary": "Blocks a specific profile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "session"
                ],
                "properties": {
                  "session": {
                    "type": "string",
                    "format": "session"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/profiles/{id}/unblock": {
      "post": {
        "summary": "Unblocks a specific profile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "session"
                ],
                "properties": {
                  "session": {
                    "type": "string",
                    "format": "session"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/posts": {
      "get": {
        "summary": "Get posts",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string"
          },
          {
            "in": "query",
            "name": "profileId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Array of posts"
          }
        }
      },
      "post": {
        "summary": "Create new post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "accountId",
                  "text"
                ],
                "properties": {
                  "accountId": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created - Post Id"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "summary": "Get post with interactions",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Post with interactinos"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted - Post Id"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/posts/{id}/like": {
      "put": {
        "summary": "Like a post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "accountId"
                ],
                "properties": {
                  "accountId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created - Post Id"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/posts/{id}/share": {
      "put": {
        "summary": "Share a post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "accountId",
                  "text",
                  "sharedPostId"
                ],
                "properties": {
                  "accountId": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created - Post Id"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/posts/{id}/reply": {
      "put": {
        "summary": "Reply to a post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "accountId",
                  "text",
                  "sharedPostId"
                ],
                "properties": {
                  "accountId": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created - Post Id"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/chats": {
      "post": {
        "summary": "Create new chat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sessionId", "accountIds"
                ],
                "properties": {
                  "sessionId": {
                    "type": "string"
                  },
                  "accountIds": {
                    "type": "array",
                    "items": [
                      "accountId1",
                      "accountId2"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK - Created"
          }
        }
      }
    },
    "/chats/{id}": {
      "post": {
        "summary": "Send Message",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sessionId",
                  "senderId",
                  "text"
                ],
                "properties": {
                  "sessionId": {
                    "type": "string"
                  },
                  "senderId": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK - Chat"
          }
        }
      }
    },
    "/chats/{id}/{session}": {
      "get": {
        "summary": "Get Chat By Id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string"
          },
          {
            "in": "path",
            "name": "session",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Chat"
          }
        }
      }
    },
    "/groups": {
      "post": {
        "summary": "Create new chat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sessionId",
                  "accountIds",
                  "ownerAccountId",
                  "isPrivate"
                ],
                "properties": {
                  "sessionId": {
                    "type": "string"
                  },
                  "accountIds": {
                    "type": "array",
                    "items": [
                      "accountId1",
                      "accountId2"
                    ]
                  },
                  "ownerAccountId": {
                    "type": "string"
                  },
                  "isPrivate": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK - Created"
          }
        }
      }
    },
    "/groups/{id}": {
      "post": {
        "summary": "Send Message",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sessionId",
                  "senderId",
                  "text"
                ],
                "properties": {
                  "sessionId": {
                    "type": "string"
                  },
                  "senderId": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK - Chat"
          }
        }
      },
      "put": {
        "summary": "Add User",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sessionId",
                  "accountId",
                  "chatToken"
                ],
                "properties": {
                  "sessionId": {
                    "type": "string"
                  },
                  "accountId": {
                    "type": "string"
                  },
                  "chatToken": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK user added"
          }
        }
      }
    },
    "/groups/{id}/{session}": {
      "get": {
        "summary": "Get Group By Id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string"
          },
          {
            "in": "path",
            "name": "session",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Group"
          }
        }
      }
    },
    "/hashtags": {
      "get": {
        "summary": "Get hashtags",
        "responses": {
          "200": {
            "description": "OK - Hashtags"
          }
        }
      }
    }
  }
}